Interview Question
To solve in O(1) complexity, and O(N) running time, used block for processig string in the NSRange, inside the block checked if NSMutableSet NOT containsObject each char apart. if it is not contained just added inside the set. The similar I used in Java with help of HashTable dictionary type.
Option I usedd NSStringEnumerationByComposedCharacterSequences